{"schema":"decodex/commit/1","summary":"Split SQLite run persistence modules by table family","authority":"manual"}
