+++ apps/decodex/src/state/sqlite_store/queries/runs/retry_budget.rs
@@ -1,0 +1,35 @@
+use crate::state::sqlite_store::{
+	SqliteStateStore,
+	queries::{self, Result},
+};
+
+impl SqliteStateStore {
+	pub(in crate::state) fn retry_budget_attempt_count(&self, issue_id: &str) -> Result<i64> {
+		self.connection
+			.query_row(
+				"SELECT COUNT(*) FROM run_attempts \
+				 WHERE issue_id = ?1 AND status IN ('failed', 'interrupted', 'terminal_guarded')",
+				queries::params![issue_id],
+				|row| row.get(0),
+			)
+			.map_err(Into::into)
+	}
+
+	pub(in crate::state) fn issue_has_retry_budget_attempt_after(
+		&self,
+		issue_id: &str,
+		attempt_number: i64,
+	) -> Result<bool> {
+		let count = self.connection.query_row(
+			"SELECT COUNT(*) FROM run_attempts \
+			 WHERE issue_id = ?1 \
+			 AND attempt_number > ?2 \
+			 AND status IN ('failed', 'interrupted', 'terminal_guarded') \
+			 LIMIT 1",
+			queries::params![issue_id, attempt_number],
+			|row| row.get::<_, i64>(0),
+		)?;
+
+		Ok(count > 0)
+	}
+}
